First implementation of logging. Added some basic logs for creating/deleting templates and adding/deleting tasks. I have not currently added anything for updating tasks.
Also added basic logging for creating/deleted checklists and for checking/unchecking tasks in a checklists. The information available can be updated soon. Will be integrating into the newsfeed soon.

The log includes: the action, name of the item the action was perfomed upon, the user who did the action, the time/date the action occured